requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
qrcode[pil]
pydantic
python-multipart